+++ server/libraries/commons/commons-stone/src/main/java/com/kamyczki/commons/stone/StoneKafkaConstants.java
@@ -1,0 +1,9 @@
+package com.kamyczki.commons.stone;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class StoneKafkaConstants {
+
+    public static String STONE_EVENTS_TOPIC = "stone_events";
+}
