+++ server/libraries/commons/commons-stone/src/main/java/com/kamyczki/commons/stone/events/StoneCreatedEventDto.java
@@ -1,0 +1,12 @@
+package com.kamyczki.commons.stone.events;
+
+import lombok.*;
+
+@Getter
+@RequiredArgsConstructor
+@ToString(callSuper = true)
+public class StoneCreatedEventDto extends StoneEventDto{
+    private String name;
+    private String description;
+    private String zipCode;
+}
