+++ server/libraries/commons/commons-stone/src/main/java/com/kamyczki/commons/stone/events/StoneEventType.java
@@ -1,0 +1,6 @@
+package com.kamyczki.commons.stone.events;
+
+public enum StoneEventType {
+    CREATED,
+    FOUND
+}
