+++ server/libraries/commons/pom.xml
@@ -14,6 +14,7 @@
 
     <modules>
         <module>commons-error</module>
+        <module>commons-stone</module>
     </modules>
 
 </project>
