+++ server/microservices/stone-read/src/main/java/com/kamyczki/stone/read/StoneMapper.java
@@ -1,0 +1,16 @@
+package com.kamyczki.stone.read;
+
+import com.kamyczki.commons.stone.events.StoneCreatedEventDto;
+import com.kamyczki.stone.read.dto.StoneDto;
+import org.mapstruct.Mapper;
+import org.mapstruct.factory.Mappers;
+
+@Mapper
+public interface StoneMapper {
+
+    StoneMapper INSTANCE = Mappers.getMapper(StoneMapper.class);
+
+    StoneDto doStoneDto(Stone stone);
+
+    Stone toStone(StoneCreatedEventDto stoneCreatedEvent);
+}
