+++ server/microservices/stone-read/src/main/java/com/kamyczki/stone/read/StoneRepository.java
@@ -1,0 +1,8 @@
+package com.kamyczki.stone.read;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StoneRepository extends JpaRepository<Stone,String> {
+}
