+++ server/microservices/stone-read/src/main/java/com/kamyczki/stone/read/dto/StoneDto.java
