+++ server/microservices/stone-read/src/main/resources/application.yml
@@ -1,0 +1,23 @@
+server:
+  port: 8082
+
+spring:
+  application:
+    name: stone
+
+  datasource:
+    username: postgres
+    password: postgres
+    url: jdbc:postgresql://localhost:5433/postgres
+    driver-class-name: org.postgresql.Driver
+  liquibase:
+    change-log: classpath:/db/changelog-master.yml
+  kafka:
+    bootstrap-servers: localhost:9200
+    consumer:
+      group-id: my-consumer-group
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      properties:
+        enable.auto.commit: false
