+++ server/microservices/stone-read/src/main/resources/db/changelog-master.yml
@@ -1,0 +1,4 @@
+databaseChangeLog:
+  - include:
+      file: ./changelogs/tb-1.xml
+      relativeToChangelogFile: true
