+++ server/microservices/stone-read/src/main/resources/docker-compose.yml
@@ -1,0 +1,17 @@
+version: '3.8'
+services:
+  db:
+    container_name: stone
+    image: postgres:14.1-alpine
+    restart: always
+    environment:
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+      - POSTGRES_DATABASE=postgres
+    ports:
+      - '5433:5432'
+    volumes:
+      - db:/var/lib/postgresql/data
+volumes:
+  db:
+    driver: local
