+++ server/microservices/stone-read/src/test/java/com/kamyczki/stone/StoneReadApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kamyczki.stone;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StoneReadApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
