+++ server/microservices/stone-write/src/main/java/com/kamyczki/stone/StoneWriteApplication.java
@@ -1,0 +1,15 @@
+package com.kamyczki.stone;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@EnableDiscoveryClient
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
+public class StoneWriteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StoneWriteApplication.class, args);
+	}
+}
