+++ server/microservices/stone-write/src/main/java/com/kamyczki/stone/write/event/StoneCreatedEvent.java
@@ -1,0 +1,14 @@
+package com.kamyczki.stone.write.event;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@ToString(callSuper = true)
+public class StoneCreatedEvent extends StoneEvent{
+    private String name;
+    private String description;
+    private String zipCode;
+}
