+++ server/microservices/stone-write/src/main/java/com/kamyczki/stone/write/event/StoneUtils.java
