--- server/microservices/stone/src/main/java/com/kamyczki/stone/StoneApplication.java
@@ -1,15 +1,0 @@
-package com.kamyczki.stone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@EnableDiscoveryClient
-@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
-public class StoneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StoneApplication.class, args);
-	}
-}
