--- server/microservices/stone/src/main/java/com/kamyczki/stone/read/StoneRepository.java
@@ -1,6 +1,0 @@
-package com.kamyczki.stone.read;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoneRepository extends JpaRepository<Stone,String> {
-}
