--- server/microservices/stone/src/main/java/com/kamyczki/stone/read/kafka/StoneEventKafkaConsumer.java
@@ -1,4 +1,0 @@
-package com.kamyczki.stone.read.kafka;
-
-public class StoneEventKafkaConsumer {
-}
