--- server/microservices/stone/src/main/java/com/kamyczki/stone/write/event/type/StoneEventType.java
@@ -1,6 +1,0 @@
-package com.kamyczki.stone.write.event.type;
-
-public enum StoneEventType {
-    CREATED,
-    FOUND
-}
