--- server/microservices/stone/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-server:
-  port: 8082
-
-spring:
-  application:
-    name: stone
-
-  data:
-    mongodb:
-      uri: mongodb+srv://natalia:JJe7UxP2HtSDaP3R@kamyczki.kjqj1.mongodb.net/stone
